fix: ensure minio creds in namespace are updated
Fixes an issue where the metacontroller defined by the kfp-profile-controller does not update the object storage credentials (stored in the `mlpipeline-minio-artifact` secret in each user's namespace), even though the kfp-profile-controller itself is informed of the new creds.  This was caused by using the `OnDelete` update mode in the metacontroller, which will only update the object if it is fully deleted (and not if it has just changed).  This is now set to `InPlace`, which always updates the object if the desired state differs from the current state (similar to if you did a `kubectl apply`).

Closes #56