feat: convert bundle tests to pytest-operator (#85)
This adds a pytest-operator driven test suite for the bundle.  Previously, the bundle was deployed "manually" in integrate.yaml.  This refactors the test suite to deploy using pytest-operator and by substituting locally built charms into a template bundle.

This commit includes some transferable helpers that should in future be migrated to a shared repo (either in pytest-operator or somewhere else like charmed-kubeflow-chisme)

Also included: 
* Adds notes to disabled test_pipelines.py