fix: add catch for when CI fails to deploy a bundle
Previously, if an error occurred during the bundle-integration tests `juju deploy ./bundle.yaml ...` call, this would not be caught and the tests would proceed as if the call succeeded.  This led to unexpected behaviour, such as sometimes tests passing even if the bundle was not deployed.  This PR adds a check on the return code and raises an error message when needed.