fix: kfp-persistence config values
A new environment variable was added to the kfp-persistence workload in the release associated with Kubeflow KF 1.6.1.  Upstream, this is added by configmap.  The previous code tried to add this as a kubernetesResource in podspec, but configMaps are not supported in that way.  This commit instead applies the environment directly with other environment variables.