tests: fix kfp-api integration tests in multiple places
The kfp-api individual tests were failing due to a number of reasons, mainly caused by how
this charm's dependencies are deployed and the time it takes for those to be up and ready.
This commit ensures those dependencies (mysql, minio, prometheus) are active and idle within
a reasonable time to avoid timeouts and charms misconfigurations.
1. For mysql-k8s, the charm is now deployed from channel 8.0/edge, a version that contains the testing
profile option, which constraints the charm in a recommended way to deploy on GitHub runners.
2. For prometheus-k8s, since this charms requires a significant amount of time to deploy, the wait
for idle timeout is increased to 90*20. This change was introduced by canonical/kfp-operators#178,
but incorrectly reverted by canonical/kfp-operators#197.

Part of #243