refactor: charmed-kubeflow-chisme's base charm to re-write kfp-schedwf (#288)
* refactor: charmed-kubeflow-chisme's base charm to re-write this kfp-schedwf

Rewrite this charm, migrating it from podspec to sidecar using the reusable base
charm in charmed-kubeflow-chisme. With this refactor a couple of files were added:
* src/componets: contains components that are used by this charm's reconciler.
More information in [1].
* src/templates: contains jinja2 template files with kubernetes resources that the
charm is in charge of applying and managing. This is of course handled by the
reconciler mentioned before.

Links:
[1] https://github.com/canonical/charmed-kubeflow-chisme/tree/add-base-charm/src/charmed_kubeflow_chisme/components#reusable-charm-components

tests: add unit tests for testing the new charm code, also add tests deps

tests: enable trust in repo level integration tests for deploying kfp-schedwf