fix: Handle relation-broken events (#272)
* api: Handle relation-broken events

In the kpf-api charm, '*-relation-broken' events trigger the same
workflow as every other event, forcing the charm to recalculate its
configuration. In order to do that, it attempts to retrieve info from
its required relations. When executing in response to a relation-broken
event for a removed relational-db relation, however, the library call to
fetch the data raises an error. We need to catch this error and properly
propagate it so that the charm lands in the desired state (i.e. Blocked,
waiting for the required relation to be added).

Closes #222

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@canonical.com>