feat: add sa token component to kfp-persistence
The kfp-persistence charms needs a ServiceAccount token to be
inside the workload container to be able to start the service.
The sa token component will generate the required token and save it
in a file where it can be accessed to later be pushed from the charm
to the workload container.

Fixes #343