Update bug_report.yml to remove shell attribute in logs
Logs are saved as shell input, which means they get rendered directly as a code block in the final output.  Let's change this to not be rendered.

The motivation for this is to enable something like [this issue's logs](https://github.com/canonical/kfp-operators/issues/367#issue-1972405752), where we can use a `<details> ... </details>` block to hide the logs in an expandable text block.  If the logs are rendered as shell, the user cannot do this.