fix: use k8s service patch to rename kfp-api svc and share correct data
Use the KubernetesServicePatch for patching the service that juju creates
for the kfp-api charm so it gets the name that other pipeline components
expect (ml-pipeline).
This commit also gets rid of the Service manifest
that was applied manually by the charm as we don't need it anymore because
we are patching the svc as we desire.
Finally, this commit also ensures the correct service name is passed to
other related applications.