fix: reduce flakiness of kfp-api integration tests
during deployment, the mysql-k8s charm sometimes lands temporarily in an error state.  This error state resolves over time and the charm continues on to work as expected, but because we do `wait_for_idle(['mysql-k8s'], ...raise_on_error=True)` the temporary error state results in a failure of our test.  This fix sets `raise_on_error=False` to avoid this.

This does not change the passing condition of the test - the test will only pass if mysql-k8s goes to Active/Idle.  The only change is that mysql-k8s going to Error state does not result in a fast failure - the test will continue until the timeout before failing.