build: install charmed-kubeflow-chisme from binary
This commit removes charmed-kubeflow-chisme from the requirements file to prevent
it to be built at charm build time, as some of its build deps are causing the
issue described in canonical/bundle-kubeflow#883.
Also in this commit the requirements-unit.in has been changed to include
charmed-kubeflow-chisme to install it during unit tests as it was
removed from the requirements.txt; therefore from the
requirements-unit.txt file.
Finally, this commit removes the "usage" legend from
requirements-unit.in as it is not a team standard anymore.

Part of canonical/bundle-kubeflow#883