ci, fix: change profile of mysql-k8s to testing, skip artefact collection on success (#624)
* ci, fix: change profile of mysql-k8s to testing

The mysql-k8s charm, used as kfp-db, has a profile configuration option that can
be set to testing which sets the scope of deployment. This is recommended for
testing environments with limited memory/disk.

* ci: skip collection of debug artefacts on success

The Github runners for the KFP CI are getting out of storage by the
time the tests execution finishes, so there is no space left for saving
logs. Skipping saving the debug artefacts on success will allow the workflows
to run successfully.